pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rtphelper"
version = "0.2.0"
description = "Web application for remote RTP capture and SRTP decryption on macOS Apple Silicon"
authors = [{ name = "RTP Support Engineering" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "boto3>=1.34",
  "click>=8.1",
  "fastapi>=0.115",
  "jinja2>=3.1",
  "pydantic>=2.8",
  "pylibsrtp>=0.10",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0",
  "scapy>=2.5.0",
  "uvicorn>=0.30"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
rtphelper = "rtphelper.cli:main"

[tool.setuptools.package-data]
rtphelper = [
  "web/templates/*.html",
  "web/static/*.css",
  "web/static/*.js",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rtphelper*"]
